requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==2.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.7
psycopg2==2.7.4
ptyprocess==0.6.0
Pygments==2.7.1
pytz==2020.1
sqlparse==0.4.1
traitlets==5.0.4
wcwidth==0.2.5
